Rename Map2 to Map, Autocomplete created as a separate component
