+++ client/src/App.js
@@ -9,7 +9,7 @@ import store from "./store";
 import NavBar from "./components/NavBar/NavBar";
 import LoginPage from "./components/LoginPage/LoginPage";
 import SignUp from "./components/RegisterPage/SignUp";
-import Map2 from "./components/Map/Map2";
+import Map from "./components/Map/Map";
 import ReportIncident from "./components/ReportIncident/ReportIncident";
 
 function App() {
@@ -18,8 +18,8 @@ function App() {
       <Router>
         <NavBar />
         <Routes>
-          <Route exact path="/" exact element={<Map2 />} />
-          <Route path="/map" element={<Map2 />} />
+          <Route exact path="/" exact element={<Map />} />
+          <Route path="/map" element={<Map />} />
           <Route path="/login" element={<LoginPage />} />
           <Route path="/signup" element={<SignUp />} />
           <Route path="/report" element={<ReportIncident />} />
